requirements.txt
pandas>=2.0
numpy>=1.24
scipy>=1.10
requests>=2.31
tqdm>=4.66
PyYAML>=6.0
